Check for directory

+++ install.sh
@@ -29,6 +29,13 @@ if [[ "$force" == "true" ]]; then
 fi
 
 destination_path="$HOME/.vscode/extensions/"
-extension_path="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
+if [ ! -d "$destination_path" ]; then
+  echo "$destination_path does not exist" >&2
+  exit 1
+fi
+extension_path="$(
+  cd "$(dirname "$0")" >/dev/null 2>&1
+  pwd -P
+)"
 rsync -a --delete${dry_run} --verbose --exclude=install.sh --exclude=.gitignore \
   --exclude=.git "$extension_path" "$destination_path"
